+++ modules/canbus/vehicle/lexus/protocol/headlight_cmd_118.cc
@@ -62,7 +62,7 @@ Headlightcmd118* Headlightcmd118::set_ignore_overrides(bool ignore_overrides) {
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Headlightcmd118::set_p_ignore_overrides(uint8_t* data,
                                              bool ignore_overrides) {
-  int x = ignore_overrides;
+  uint8_t x = ignore_overrides;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 1, 1);
@@ -77,7 +77,7 @@ Headlightcmd118* Headlightcmd118::set_enable(bool enable) {
 // 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 0, 'type': 'bool',
 // 'order': 'motorola', 'physical_unit': ''}
 void Headlightcmd118::set_p_enable(uint8_t* data, bool enable) {
-  int x = enable;
+  uint8_t x = enable;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 0, 1);
@@ -92,7 +92,7 @@ Headlightcmd118* Headlightcmd118::set_clear_override(bool clear_override) {
 // 'len': 1, 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 2,
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Headlightcmd118::set_p_clear_override(uint8_t* data, bool clear_override) {
-  int x = clear_override;
+  uint8_t x = clear_override;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 2, 1);
@@ -107,7 +107,7 @@ Headlightcmd118* Headlightcmd118::set_clear_faults(bool clear_faults) {
 // 'len': 1, 'is_signed_var': False, 'physical_range': '[0|1]', 'bit': 3,
 // 'type': 'bool', 'order': 'motorola', 'physical_unit': ''}
 void Headlightcmd118::set_p_clear_faults(uint8_t* data, bool clear_faults) {
-  int x = clear_faults;
+  uint8_t x = clear_faults;
 
   Byte to_set(data + 0);
   to_set.set_value(x, 3, 1);
@@ -126,7 +126,7 @@ Headlightcmd118* Headlightcmd118::set_headlight_cmd(
 // 'order': 'motorola', 'physical_unit': ''}
 void Headlightcmd118::set_p_headlight_cmd(
     uint8_t* data, Headlight_cmd_118::Headlight_cmdType headlight_cmd) {
-  int x = headlight_cmd;
+  uint8_t x = headlight_cmd;
 
   Byte to_set(data + 1);
   to_set.set_value(x, 0, 8);
